pyproject.toml
[project]
name = "utility-explorer"
version = "0.1.0"
description = "Personal productivity tools for keeping balls in the air"
requires-python = ">=3.10"
dependencies = [
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.108.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "anthropic>=0.39.0",
]

[project.scripts]
ue = "ue.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
